test(talk): consolidate relay ownership coverage
